Changed views file to package
