+++ company_profiles_app/views/edit_company.py
@@ -1,0 +1,7 @@
+# from django.views import View
+#
+#
+# class EditCompanyView(View):
+#     def get(self, request, *args, **kwargs):
+#
+#
